+++ packages/backend/src/backend_factory.ts
@@ -18,7 +18,7 @@ import {
   BranchBackendIdentifier,
   SandboxBackendIdentifier,
 } from '@aws-amplify/platform-core';
-import { stackOutputKey } from '@aws-amplify/backend-output-schemas';
+import { platformOutputKey } from '@aws-amplify/backend-output-schemas';
 import { fileURLToPath } from 'url';
 import { Backend } from './backend.js';
 import { AmplifyConsoleLinkerConstruct } from './engine/console-linker/amplify_console_linker_construct.js';
@@ -62,13 +62,14 @@ export class BackendFactory<
     );
 
     const uniqueBackendIdentifier = getUniqueBackendIdentifier(stack);
-    outputStorageStrategy.addBackendOutputEntry(stackOutputKey, {
+    outputStorageStrategy.addBackendOutputEntry(platformOutputKey, {
       version: '1',
       payload: {
         deploymentType:
           uniqueBackendIdentifier instanceof SandboxBackendIdentifier
             ? BackendDeploymentType.SANDBOX
             : BackendDeploymentType.BRANCH,
+        region: stack.region,
       },
     });