+++ packages/sandbox/src/file_watching_sandbox.test.ts
@@ -717,6 +717,20 @@ void describe('Sandbox using local project name resolver', () => {
 
     assert.equal(mockListener.mock.callCount(), 1);
   });
+
+  void it('emits the successfulDeletion event after delete is finished', async () => {
+    const mockListener = mock.fn();
+    ({ sandboxInstance, fileChangeEventCallback } = await setupAndStartSandbox({
+      executor: sandboxExecutor,
+      cfnClient: cfnClientMock,
+    }));
+
+    sandboxInstance.on('successfulDeletion', mockListener);
+
+    await sandboxInstance.delete({});
+
+    assert.equal(mockListener.mock.callCount(), 1);
+  });
 });
 
 /**