+++ packages/auth-construct/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-amplify/auth-construct-alpha
 
+## 0.2.1
+
+### Patch Changes
+
+- d0119b25: Flatten loginWith type to improve autocompletion.
+
 ## 0.2.0
 
 ### Minor Changes